+++ torch_geometric/transforms/delaunay.py
@@ -8,7 +8,7 @@ from torch_geometric.transforms import BaseTransform
 
 
 class _QhullTransform(BaseTransform):
-    r"""Q-hull implementation of Delaunay triangulation, uses NumPy arrays."""
+    r"""Q-hull implementation of delaunay triangulation."""
     def forward(self, data: Data) -> Data:
         assert data.pos is not None
         import scipy.spatial
@@ -22,9 +22,7 @@ class _QhullTransform(BaseTransform):
 
 
 class _ShullTransform(BaseTransform):
-    r"""Sweep-hull implementation of Delaunay triangulation, uses Torch
-    tensors.
-    """
+    r"""Sweep-hull implementation of delaunay triangulation."""
     def forward(self, data: Data) -> Data:
         assert data.pos is not None
         from torch_delaunay.functional import shull2d
@@ -35,22 +33,19 @@ class _ShullTransform(BaseTransform):
 
 
 class _SequentialTransform(BaseTransform):
-    r"""Runs the specifies transforms one after another sequentially.
+    r"""Runs the first successful transformation.
 
-    Returns the first successful result, and raises exception otherwise. All
-    intermediate exceptions are suppressed except the last.
+    All intermediate exceptions are suppressed except the last.
     """
     def __init__(self, transforms: List[BaseTransform]) -> None:
-        if (num_transforms := len(transforms)) == 0:
-            raise ValueError(
-                f"at least one transform is expected, got {num_transforms}")
+        assert len(transforms) > 0
         self.transforms = transforms
 
     def forward(self, data: Data) -> Data:
         for i, transform in enumerate(self.transforms):
             try:
                 return transform.forward(data)
-            except ModuleNotFoundError as e:
+            except Exception as e:
                 if i == len(self.transforms) - 1:
                     raise e
         return data
@@ -62,26 +57,28 @@ class Delaunay(BaseTransform):
     (functional name: :obj:`delaunay`).
 
     .. hint::
-        Consider installing `torch_delaunay` python package to speed up
-        computations of Delaunay tessellations.
+        Consider installing
+        `torch_delaunay <https://github.com/ybubnov/torch_delaunay>`_ package
+        to speed up computation.
     """
+    def __init__(self) -> None:
+        self._transform = _SequentialTransform([
+            _ShullTransform(),
+            _QhullTransform(),
+        ])
+
     def forward(self, data: Data) -> Data:
         assert data.pos is not None
+        device = data.pos.device
 
         if data.pos.size(0) < 2:
-            data.edge_index = torch.tensor([], dtype=torch.long,
-                                           device=data.pos.device).view(2, 0)
-        if data.pos.size(0) == 2:
-            data.edge_index = torch.tensor([[0, 1], [1, 0]], dtype=torch.long,
-                                           device=data.pos.device)
+            data.edge_index = torch.empty(2, 0, dtype=torch.long,
+                                          device=device)
+        elif data.pos.size(0) == 2:
+            data.edge_index = torch.tensor([[0, 1], [1, 0]], device=device)
         elif data.pos.size(0) == 3:
-            data.face = torch.tensor([[0], [1], [2]], dtype=torch.long,
-                                     device=data.pos.device)
-        if data.pos.size(0) > 3:
-            transform = _SequentialTransform([
-                _QhullTransform(),
-                _ShullTransform(),
-            ])
-            data = transform.forward(data)
+            data.face = torch.tensor([[0], [1], [2]], device=device)
+        else:
+            data = self._transform.forward(data)
 
         return data